+++ packages/interweave-emoji-picker/CHANGELOG.md
@@ -1,4 +1,4 @@
-# 1.3.0
+# 1.3.0 - 10/23/17
 #### 🚀 New
 * Added a `disableGroups` prop to `Picker`, which will hide the tabs and disable group headers.
   * Emojis will now be grouped into a "none" group.